refac: mv PatchSpecs to MultImageStack; rename data_reader.py to multi_image_stack.py
