+++ src/careamics/dataset_ng/data_reader/__init__.py
@@ -1,5 +1,6 @@
 __all__ = [
     "MultiImageStack",
+    "PatchSpecs",
 ]
 
-from .data_reader import MultiImageStack
+from .multi_image_stack import MultiImageStack, PatchSpecs
